Delete specs for overlay decoration behavior
This is now tested in core and due to some vagaries of React’s batching
strategy combined with synchronous editor updates, the position of the
overlay is getting calculated before the autocomplete’s mini editor
has a chance to get a height, causing spurious test failure. Since this
is tested in core, the easiest path forward is just to remove the specs
for now. This will be easier to test once React is removed from the
text editor.

Signed-off-by: Nathan Sobo <nathan@github.com>
Signed-off-by: Max Brunsfeld <maxbrunsfeld@gmail.com>